+++ RELEASES.md
@@ -1,5 +1,15 @@
 # Confidential Match Source Code Releases
 
+## May 16, 2025
+
+* Version: v1.150.0
+* Lookup Service
+  * Image Digest: sha256:e8574efb38f773ccee14ba628963737b708ed7995a9882a44cbcfd78bf74e151
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:5c99d0d5c78b8656b576e915e6083a4ae2fb3c030831cb3db2f8c5924d402782)
+* Match Request Processor
+  * Image Digest: sha256:2248f5d6c3da2ffaf73678091acdb4a14bbab88a625584884aad172032e3b4ef
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:e54d3d5f3011ca1da041f2ebed012ab02b122eed19bbfdcff7b46c82fdb7e732)
+
 ## May 1, 2025
 
 * Version: v1.135.0